requirements.txt
asgiref==3.3.1
Django>=3.0,<4.0
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
style==1.1.0
update==0.0.1
djangorestframework==3.12.2